# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
...
  /v1/groups/integrations:
    get:
      security:
        - Bearer: []
      tags:
        - Group
      summary: Get Group Integrations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.GroupIntegrationsOut"
    put:
      security:
        - Bearer: []
      description: 'Owner-only. Body = types.GroupIntegrations. Secret fields use
        write-only sentinel semantics: incoming "[REDACTED]" keeps the stored
        value, "" clears it, anything else replaces it.'
      tags:
        - Group
      summary: Update Group Integrations
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/types.GroupIntegrations"
        description: Group Integrations
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.GroupIntegrationsOut"
  /v1/groups/integrations/test-ai:
    post:
      security:
        - Bearer: []
      description: Owner-only. Resolves the effective AI config and attempts a real
        round-trip against the configured provider using a fixed small test
        image, under a 30s timeout. Always 200; ok/detail report the outcome.
        detail is a sanitized, coarse string — never the raw provider error.
      tags:
        - Group
      summary: Test Group AI Integration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.TestConnectionResponse"
  /v1/groups/integrations/test-barcode:
    post:
      security:
        - Bearer: []
      description: Owner-only. Resolves the effective barcode config and attempts a
        real lookup against barcodespider.com for a fixed, stable EAN. Always
        200; ok/detail report the outcome. detail is a sanitized, coarse string
        — never the raw error.
      tags:
        - Group
      summary: Test Group Barcode Integration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.TestConnectionResponse"
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    types.GroupIntegrations:
      type: object
      properties:
        aiApiKey:
          type: string
        aiBaseUrl:
          type: string
        aiModel:
          type: string
        aiProvider:
          description: '"" inherit env | "disabled" force off | "openai_compatible" |
            "anthropic"'
          type: string
        barcodeTokenBarcodespider:
          type: string
        openFoodFactsContact:
          type: string
...
    v1.GroupIntegrationsOut:
      type: object
      properties:
        aiApiKey:
          type: string
        aiBaseUrl:
          type: string
        aiConfigured:
          type: boolean
        aiModel:
          type: string
        aiProvider:
          description: '"" inherit env | "disabled" force off | "openai_compatible" |
            "anthropic"'
          type: string
        barcodeTokenBarcodespider:
          type: string
        barcodespiderConfigured:
          type: boolean
        envAiBaseUrl:
          type: string
        envAiModel:
          type: string
        envAiProvider:
          type: string
        isOwner:
          type: boolean
        openFoodFactsContact:
          type: string
...
    v1.TestConnectionResponse:
      type: object
      properties:
        detail:
          type: string
        ok:
          type: boolean
...